Fix WebDAV mounts from the Windows WebDAV mini-redirector client implementation:
DAV: namespaced elements in a PROPFIND responses can not use a default namespace; each element must be explicitly prefixed with a namespace prefix defined in itself or a parent. If a default namespace is used the client will give no errors, but will not display any resources.

The only mention of this issue was found in:

http://www.greenbytes.de/tech/webdav/webdav-redirector-list.html

Also, fix the display of the top-level directory name when deployed as the root context of the domain.